+++ install-helm.sh
@@ -8,14 +8,14 @@ echo "***************************************************"
 
 set -uxo pipefail
 
+if command -v helm &> /dev/null
+then
+    echo "helm already installed, skipping"
+    exit 0
+fi
+
 # Install Helm [https://helm.sh/docs/intro/install/#from-script]
 curl -fsSL -o get_helm.sh https://raw.githubusercontent.com/helm/helm/main/scripts/get-helm-3
 chmod +x get_helm.sh
 sudo ./get_helm.sh
 rm get_helm.sh
-
-
-
-
-
-
